em-analysis: Extend EM v2 API to store 16-bit flags

+++ software/src/modules/em_energy_analysis/em_energy_analysis.h
@@ -50,9 +50,9 @@ private:
     void history_wallbox_daily_response(IChunkedResponse *response, Ownership *response_ownership, uint32_t response_owner_id);
     void history_energy_manager_5min_response(IChunkedResponse *response, Ownership *response_ownership, uint32_t response_owner_id);
     void history_energy_manager_daily_response(IChunkedResponse *response, Ownership *response_ownership, uint32_t response_owner_id);
-    bool set_wallbox_5min_data_point(const struct tm *utc, const struct tm *local, uint32_t uid, uint8_t flags, uint16_t power /* W */);
+    bool set_wallbox_5min_data_point(const struct tm *utc, const struct tm *local, uint32_t uid, uint16_t flags, uint16_t power /* W */);
     bool set_wallbox_daily_data_point(const struct tm *local, uint32_t uid, uint32_t energy /* daWh */);
-    bool set_energy_manager_5min_data_point(const struct tm *utc, const struct tm *local, uint8_t flags, const int32_t power[7] /* W */);
+    bool set_energy_manager_5min_data_point(const struct tm *utc, const struct tm *local, uint16_t flags, const int32_t power[7] /* W */);
     bool set_energy_manager_daily_data_point(const struct tm *local, const uint32_t energy_import[7] /* daWh */, const uint32_t energy_export[7] /* daWh */);
 
     std::list<std::function<bool(void)>> pending_data_points;
